+++ packages/orq-rc/src/models/components/partdoneevent.ts
@@ -79,7 +79,7 @@ export const PartReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01kf5nv5mdj40hc1wk5akyb295"),
+  _id: z.string().default("reasoning_01kf9vkzq6381yzpga090b8821"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$inboundSchema,
   reasoning: z.string(),
@@ -104,7 +104,7 @@ export const PartReasoningPart$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   PartReasoningPart
 > = z.object({
-  id: z.string().default("reasoning_01kf5nv5mdj40hc1wk5akyb295"),
+  id: z.string().default("reasoning_01kf9vkzq6381yzpga090b8821"),
   metadata: z.record(z.any()).optional(),
   kind: PartKind$outboundSchema,
   reasoning: z.string(),