+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -174,7 +174,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-01-17T09:50:54.473Z",
+    "2026-01-19T00:48:48.487Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -213,7 +213,7 @@ export const RetrieveDatasetResponseBody$outboundSchema: z.ZodType<
   createdById: z.string().optional(),
   updatedById: z.string().optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2026-01-17T09:50:54.473Z"))
+  updated: z.date().default(() => new Date("2026-01-19T00:48:48.487Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {