+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4703,7 +4703,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KF5NV5WSTJKBR7V5KBZYQPDR"),
+  id: z.string().default("01KF9VKZY1R2J8AB90ZT49NDM7"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>
@@ -4724,7 +4724,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KF5NV5WSTJKBR7V5KBZYQPDR"),
+  id: z.string().default("01KF9VKZY1R2J8AB90ZT49NDM7"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>