+++ packages/orq-rc/RELEASES.md
@@ -5909,3 +5909,13 @@ Based on:
 - [typescript v4.2.0-rc.37] packages/orq-rc
 ### Releases
 - [NPM v4.2.0-rc.37] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.37 - packages/orq-rc
+
+## 2026-01-19 00:51:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.688.0 (2.796.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.2.0-rc.38] packages/orq-rc
+### Releases
+- [NPM v4.2.0-rc.38] https://www.npmjs.com/package/@orq-ai/node/v/4.2.0-rc.38 - packages/orq-rc